+++ NEWS.md
@@ -1,3 +1,3 @@
 # carbcontpred 0.0.0.9000
 
-* This is the first release of carbcontpred.
+* This is the development version of carbcontpred.